Working on merging protobuf and latest.

+++ server-core/src/main/scala/com/convergencelabs/server/domain/model/OpenModelMetaData.scala
@@ -5,6 +5,6 @@ import java.time.Instant
 case class OpenModelMetaData(
   id: String,
   collection: String,
-  version: Long,
+  version: Int,
   createdTime: Instant,
   modifiedTime: Instant)
